64bit application.

+++ MameBake3D/ColiIDDlg.cpp
@@ -183,7 +183,7 @@ LRESULT CColiIDDlg::OnAdd(WORD wNotifyCode, WORD wID, HWND hWndCtl, BOOL& bHandl
 }
 LRESULT CColiIDDlg::OnDel(WORD wNotifyCode, WORD wID, HWND hWndCtl, BOOL& bHandled)
 {
-	int idnum = m_coliids.size();
+	int idnum = (int)m_coliids.size();
 
 	if( idnum <= 0 ){
 		return 0;
